requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.5
dataclasses-json==0.6.1
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.104.1
firebase-admin==6.2.0
frozenlist==1.4.0
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud==0.34.0
google-cloud-aiplatform==1.35.0
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
ipdb==0.13.13
ipython==8.16.1
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.323
langsmith==0.0.52
marshmallow==3.20.1
matplotlib-inline==0.1.6
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
stack-data==0.6.3
starlette==0.27.0
tenacity==8.2.3
tomli==2.0.1
tqdm==4.66.1
traitlets==5.12.0
typing-inspect==0.9.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.23.2
vertexai==0.0.1
wcwidth==0.2.8
yarl==1.9.2
